added/updated more data

+++ stencils/3D/1r/heterogeneous/star/variable/double/HaswellEP_E5-2695v3_CoD/stencil.c
@@ -2,9 +2,9 @@ double a[M][N][P];
 double b[M][N][P];
 double W[7][M][N][P];
 
-for(int k=1; k < M-1; ++k){
-for(int j=1; j < N-1; ++j){
-for(int i=1; i < P-1; ++i){
+for(long k=1; k < M-1; ++k){
+for(long j=1; j < N-1; ++j){
+for(long i=1; i < P-1; ++i){
 b[k][j][i] = W[0][k][j][i] * a[k][j][i]
 + W[1][k][j][i] * a[k][j][i-1] + W[2][k][j][i] * a[k][j][i+1]
 + W[3][k][j][i] * a[k-1][j][i] + W[4][k][j][i] * a[k+1][j][i]
